% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aweSOM.R
\name{aweSOM}
\alias{aweSOM}
\title{aweSOM interface}
\usage{
aweSOM()
}
\description{
Launches the (offline) web-based interface for training and visualizing
self-organizing maps.
}
\references{
Kohonen T. (2001) \emph{Self-Organizing Maps}, 3rd edition,
  Springer Press, Berlin.
}
