% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_panel_functions.R
\name{aweSOMabstraction}
\alias{aweSOMabstraction}
\title{Abstraction plot}
\usage{
aweSOMabstraction(ok.som, dat, cutoff = 0, pal = "Set3", reversePal = F)
}
\arguments{
\item{ok.som}{SOM data object}

\item{dat}{data object}

\item{cutoff}{}

\item{pal}{}

\item{reversePal}{}
}
\value{

}
\description{
Abstraction plot
}
\examples{
ok.data <- iris
## Build training data
dat <- ok.data[,c("Sepal.Length", "Sepal.Width",  "Petal.Length", "Petal.Width" )]
### Scale training data
dat <- scale(dat)
## Train SOM
### RNG Seed (for reproducibility)
### Initialization (PCA grid)
init <- aweSOM::somInit(dat, 4, 4)
ok.som <- kohonen::som(dat, grid = kohonen::somgrid(4, 4, 'hexagonal'), rlen = 100, alpha = c(0.05, 0.01), radius = c(6.08,-6.08), init = init, dist.fcts = 'sumofsquares')
aweSOMabstraction(ok.som, dat, cutoff = 0)
}
