% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_panel_functions.R
\name{aweSOMplot}
\alias{aweSOMplot}
\title{SOM interactive visualizations}
\usage{
aweSOMplot(
  ok.som,
  ok.sc = NULL,
  ok.data,
  omitRows = NULL,
  graphType = "Hitmap",
  plotNames = "(rownames)",
  plotVarMult = NULL,
  plotVarOne = NULL,
  plotOutliers = T,
  plotEqualSize = F,
  contrast = "contrast",
  average_format = "mean",
  plotSize = 100,
  palsc = "Set3",
  palplot = "viridis",
  plotRevPal = F
)
}
\arguments{
\item{ok.som}{SOM data object}

\item{ok.sc}{superclasses}

\item{ok.data}{SOM training dataset}

\item{omitRows}{}

\item{graphType}{Graph type}

\item{plotNames}{}

\item{plotVarMult}{}

\item{plotVarOne}{}

\item{plotOutliers}{}

\item{plotEqualSize}{}

\item{contrast}{}

\item{average_format}{}

\item{plotSize}{}

\item{palsc}{}

\item{palplot}{}

\item{plotRevPal}{}
}
\value{

}
\description{
SOM interactive visualizations
}
\examples{
ok.data <- iris
## Build training data
dat <- ok.data[,c("Sepal.Length", "Sepal.Width",  "Petal.Length", "Petal.Width" )]
### Scale training data
dat <- scale(dat)
## Train SOM
### RNG Seed (for reproducibility)
### Initialization (PCA grid)
init <- aweSOM::somInit(dat, 4, 4)
ok.som <- kohonen::som(dat, grid = kohonen::somgrid(4, 4, 'hexagonal'), rlen = 100, alpha = c(0.05, 0.01), radius = c(6.08,-6.08), init = init, dist.fcts = 'sumofsquares')
## Group cells into superclasses (PAM clustering)
superclust <- cluster::pam(ok.som$codes[[1]], 2)
superclasses <- unname(superclust$clustering)
variables <- c("Sepal.Length", "Sepal.Width",  "Petal.Length", "Petal.Width")
#Hitmap
aweSOM::aweSOMplot(ok.som = ok.som, ok.sc = superclasses, ok.data = ok.data, graphType = 'Hitmap', plotSize = 100)
#Radar
aweSOM::aweSOMplot(ok.som = ok.som, ok.sc = superclasses, ok.data = ok.data, graphType = 'Radar', plotVarMult = variables, plotSize = 100)
#Barplot
aweSOM::aweSOMplot(ok.som = ok.som, ok.sc = superclasses, ok.data = ok.data,  graphType = 'Barplot',  plotVarMult = variables,  plotSize = 100)
#Boxplot
aweSOM::aweSOMplot(ok.som = ok.som, ok.sc = superclasses, ok.data = ok.data, graphType = 'Boxplot', plotVarMult = variables, plotSize = 100)
#Lineplot
aweSOM::aweSOMplot(ok.som = ok.som, ok.sc = superclasses, ok.data = ok.data, graphType = 'Line', plotVarMult = variables, plotSize = 100)
}
