% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_panel_functions.R
\name{aweSOMscreeplot}
\alias{aweSOMscreeplot}
\title{Plot screeplot for clustering of superclasses}
\usage{
aweSOMscreeplot(
  ok.som,
  nclass = 2,
  method = "hierarchical",
  hmethod = "ward.D2"
)
}
\arguments{
\item{ok.som}{SOM data object}

\item{ok.hclust}{hierarchical clustering object}

\item{input_kohSuperclass}{number of superclasses}
}
\value{
Screeplot of superclass clustering
}
\description{
Plot screeplot for clustering of superclasses
}
\examples{
ok.data <- iris
## Build training data
dat <- ok.data[,c("Sepal.Length", "Sepal.Width",  "Petal.Length", "Petal.Width" )]
### Scale training data
dat <- scale(dat)
## Train SOM
### RNG Seed (for reproducibility)
### Initialization (PCA grid)
init <- aweSOM::somInit(dat, 4, 4)
ok.som <- kohonen::som(dat, grid = kohonen::somgrid(4, 4, 'hexagonal'), rlen = 100, alpha = c(0.05, 0.01), radius = c(6.08,-6.08), init = init, dist.fcts = 'sumofsquares')
## Group cells into superclasses (PAM clustering)
superclust <- cluster::pam(ok.som$codes[[1]], 2)
superclasses <- unname(superclust$clustering)
aweSOM::aweSOMscreeplot(ok.som, method = 'hierarchical', hmethod = 'ward.D2', nclass = 2)
}
