% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_panel_functions.R
\name{aweSOMsilhouette}
\alias{aweSOMsilhouette}
\title{Silhouette plot}
\usage{
aweSOMsilhouette(ok.som, ok.sc)
}
\arguments{
\item{ok.som}{SOM data object}

\item{ok.sc}{}
}
\value{
Silhouette plot for superclass clustering
}
\description{
Silhouette plot
}
\examples{
#' ok.data <- iris
## Build training data
dat <- ok.data[,c("Sepal.Length", "Sepal.Width",  "Petal.Length", "Petal.Width" )]
### Scale training data
dat <- scale(dat)
## Train SOM
### RNG Seed (for reproducibility)
### Initialization (PCA grid)
init <- aweSOM::somInit(dat,4, 4)
ok.som <- kohonen::som(dat, grid = kohonen::somgrid(4, 4, 'hexagonal'), rlen = 100, alpha = c(0.05, 0.01), radius = c(6.08,-6.08), init = init, dist.fcts = 'sumofsquares')
## Group cells into superclasses (PAM clustering)
superclust <- cluster::pam(ok.som$codes[[1]], 2)
superclasses <- unname(superclust$clustering)
aweSOM::aweSOMsilhouette(ok.som, superclasses)
}
