% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_panel_functions.R
\name{aweSOMsmoothdist}
\alias{aweSOMsmoothdist}
\title{Smooth distance plot}
\usage{
aweSOMsmoothdist(x, pal = "viridis", reversePal = F)
}
\arguments{
\item{ok.som}{SOM data object}

\item{ok.dist}{}

\item{input_palplot}{}

\item{input_plotRevPal}{}
}
\value{

}
\description{
Smooth distance plot
}
