% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_panel_functions.R
\name{getPalette}
\alias{getPalette}
\title{Generate plot color palets
Function that generates color pallets}
\usage{
getPalette(pal, n, reverse = F)
}
\arguments{
\item{pal}{}

\item{n}{}

\item{reverse}{}
}
\value{

}
\description{
Generate plot color palets
Function that generates color pallets
}
