% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_panel_functions.R
\name{getPlotParams}
\alias{getPlotParams}
\title{D3-based plot data generation
Function generates a list object containing data for D3-based plots}
\usage{
getPlotParams(
  type,
  som,
  superclass,
  data,
  plotsize,
  varnames,
  normtype = c("range", "contrast"),
  palsc,
  palplot,
  cellNames,
  plotOutliers,
  reversePal,
  options = NULL,
  valueFormat
)
}
\arguments{
\item{type}{}

\item{som}{}

\item{superclass}{}

\item{data}{}

\item{plotsize}{}

\item{varnames}{}

\item{normtype}{}

\item{palsc}{}

\item{palplot}{}

\item{cellNames}{}

\item{plotOutliers}{}

\item{reversePal}{}

\item{options}{}

\item{valueFormat}{}
}
\value{

}
\description{
D3-based plot data generation
Function generates a list object containing data for D3-based plots
}
