% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_panel_functions.R
\name{ok.som.function}
\alias{ok.som.function}
\title{Train SOM}
\usage{
ok.som.function(
  ok.traindat,
  input_trainSeed,
  input_kohInit,
  input_kohDimy,
  input_kohDimx,
  input_kohTopo,
  input_trainRlen,
  input_trainAlpha1,
  input_trainAlpha2,
  input_trainRadius1,
  input_trainRadius2
)
}
\arguments{
\item{ok.traindat}{return object from ok.traindat.function}

\item{input_trainSeed}{random seed}

\item{input_kohInit}{"pca" for PCA initialization}

\item{input_kohDimy}{integer indicating y dimension of SOM}

\item{input_kohDimx}{integer indicating x dimension of SOM}

\item{input_kohTopo}{"hexagonal" or "recantgular" for geometrical structure of SOM}

\item{input_trainRlen}{integer. 100?}

\item{input_trainAlpha1}{integer/double}

\item{input_trainAlpha2}{integer/double}

\item{input_trainRadius1}{integer/double}

\item{input_trainRadius2}{integer/double}
}
\value{
res object
}
\description{
Train SOM
}
\examples{
ok.som.function(ok.traindat = ok.traindat, input_trainSeed = 61929, 
                          input_KohInit = "pca", input_kohDimy= 10, input_kohDimx = 10,
                          input_kohTopo = "hexagonal", input_trainRlen = 100,
                          input_trainAlpha1 = 0.05, input_trainAlpha2 = 0.01, 
                          input_trainRadius1 = 6.08, input_trainRadius2 = -6.08,)
}
