% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_panel_functions.R
\name{ok.traindat.function}
\alias{ok.traindat.function}
\title{Title}
\usage{
ok.traindat.function(input_trainscale, ok.data, varSelected, varWeights)
}
\arguments{
\item{input_trainscale}{i.e. an checkboxInput}

\item{ok.data}{i.e. the data object storing the data frame  ok.data object}

\item{varSelected}{i.e. selected variables to train SOM on, vector with TRUE/FALSE values}

\item{varWeights}{i.e. the weights assigned to these variables (format vector of length of variables no.)
when no weights assigned c(1,1,1,1,...)}

\item{input_trainbutton}{i.e.  an action button}
}
\value{

}
\description{
Title
}
