% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_panel_functions.R
\name{somDist}
\alias{somDist}
\title{Distance measures}
\usage{
somDist(x)
}
\arguments{
\item{x}{```kohonen``` object, a SOM created by the ```som``` function.}
}
\value{
list with distance measures
}
\description{
Distance measures
}
\examples{
somDist(x = ok.som)
}
