% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_panel_functions.R
\name{somQuality}
\alias{somQuality}
\title{Calculate SOM quality measures}
\usage{
somQuality(ok.som, traindat)
}
\arguments{
\item{ok.som}{SOM object created by ok.som.function}

\item{traindat}{matrix containing the training data}
}
\value{
list with quality measures
}
\description{
Calculate SOM quality measures
}
\examples{
somQuality(ok.som  = ok.som, traindat = traindat)

}
