% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_panel_functions.R
\name{aweSOMplot}
\alias{aweSOMplot}
\title{SOM interactive visualizations}
\usage{
aweSOMplot(
  ok.som,
  ok.sc = NULL,
  ok.data,
  omitRows = NULL,
  graphType = "Hitmap",
  plotNames = "(rownames)",
  plotVarMult = NULL,
  plotVarOne = NULL,
  plotOutliers = T,
  plotEqualSize = F,
  contrast = "contrast",
  average_format = "mean",
  plotSize = 100,
  palsc = "Set3",
  palplot = "viridis",
  plotRevPal = F,
  elementId = NULL
)
}
\arguments{
\item{ok.som}{SOM data object}

\item{ok.sc}{superclasses}

\item{ok.data}{SOM training dataset}

\item{omitRows}{Select to omit specific rows in the ok.data argument.}

\item{graphType}{The graph type. Either "hitmap", "radar", "barplot", "boxplot" ,or "lineplot". Default is "hitmap"}

\item{plotNames}{Select variable by which observations per each cell are displayed as you hover over a cell. Default is "(rownames)"}

\item{plotVarMult}{Plotted variables. Concerning the radar, barplot, boxplot, lineplot, this argument allows selecting the variables
which should be displayed within the SOM visualization as a vector of characters indicating the names of the variables. Concering
visualizations of categorical variables, only one variable can be passed to the functon.}

\item{plotVarOne}{Julien you probabyl will mege this one so I leave it blank}

\item{plotOutliers}{Indicating whether outlier variables in the "boxplot" SOM visualization are displayed. Default is TRUE (outliers are displayed)}

\item{plotEqualSize}{???}

\item{contrast}{Controls the scaling of the variables. Default is "constrast" which scales each of the variables indepenently. 
Alternatively, same scales can be used where all variables are displayed on the identical scaled based on global minimum and 
maximum values by using "no_constrast". Or the observations range by using "range"}

\item{average_format}{The type of average used for the visualizated variables. Default option is "mean", alternatively 
"median" can be used to displayed the median value of the variables or "prototype" which is the value of the SOM cell protoype}

\item{plotSize}{Plot size of the SOM visualization measured as integer value in pixel. Default is 100.}

\item{palsc}{The color palette for visualizing superclasses of SOM. Default is "Set3"}

\item{palplot}{The  color palette for visualizing variables in SOM cells. Default is "viridis"}

\item{plotRevPal}{Boolean whether color palette for variables is reversed. Default is FALSE}

\item{elementId}{(optional) user-defined elementId of the widget}
}
\value{
Interactive html-widget visualizing a SOM.
}
\description{
This function allows to create interactive visualizations of self-organizing maps (SOM) as html-widgets. 
The following types of visualizations can be displayed that reflect upon distribution of observations per cell and 
distributions of variables per cell: hitmap, radar, barplot, boxplot, lineplot. Interactive behavior enables
the display of the respective observations per cell as well as further statistical information on 
selected variables when hovering over these.
}
\examples{
## Build training data
dat <- iris[, c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")]
### Scale training data
dat <- scale(dat)
## Train SOM
### RNG Seed (for reproducibility)
### Initialization (PCA grid)
init <- aweSOM::somInit(dat, 4, 4)
ok.som <- kohonen::som(dat, grid = kohonen::somgrid(4, 4, 'hexagonal'), 
                       rlen = 100, alpha = c(0.05, 0.01), 
                       radius = c(6.08,-6.08), init = init, 
                       dist.fcts = 'sumofsquares')
## Group cells into superclasses (PAM clustering)
superclust <- cluster::pam(ok.som$codes[[1]], 2)
superclasses <- unname(superclust$clustering)

variables <- c("Sepal.Length", "Sepal.Width",  "Petal.Length", "Petal.Width")
#Hitmap
aweSOM::aweSOMplot(ok.som = ok.som, ok.sc = superclasses, ok.data = iris, 
                   graphType = 'Hitmap', plotSize = 100)
#Radar
aweSOM::aweSOMplot(ok.som = ok.som, ok.sc = superclasses, ok.data = iris, 
                   graphType = 'Radar', plotVarMult = variables, 
                   plotSize = 100)
#Barplot
aweSOM::aweSOMplot(ok.som = ok.som, ok.sc = superclasses, ok.data = iris,
                   graphType = 'Barplot', plotVarMult = variables,
                    plotSize = 100)
}
