% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_panel_functions.R
\name{aweSOMscreeplot}
\alias{aweSOMscreeplot}
\title{Screeplot}
\usage{
aweSOMscreeplot(
  ok.som,
  nclass = 2,
  method = "hierarchical",
  hmethod = "ward.D2"
)
}
\arguments{
\item{ok.som}{SOM data object}

\item{nclass}{number of superclasses to be visualized in the screeplot. Default is 2.}

\item{method}{Method used for clustering. Hierarchical clustering ("hierarchical") and PAM ("pam") clustering can be used. 
By default hierarchical clustering is applied.}

\item{hmethod}{Specifically for hierarchicical clustering, the clustering method can be specified. By default "ward.D2" is used
for other options, see the stats::hclust method documentation.}
}
\value{
Screeplot
}
\description{
Creates a screeplot that that provides a quality measurement of the quality of the clustering of each SOM cell 
for both PAM and hierarchical clustering.
}
\examples{
## Build training data
dat <- iris[, c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")]
### Scale training data
dat <- scale(dat)
## Train SOM
### RNG Seed (for reproducibility)
### Initialization (PCA grid)
init <- somInit(dat, 4, 4)
ok.som <- kohonen::som(dat, grid = kohonen::somgrid(4, 4, 'hexagonal'), 
                       rlen = 100, alpha = c(0.05, 0.01), 
                       radius = c(6.08,-6.08), 
                       init = init, dist.fcts = 'sumofsquares')
## Group cells into superclasses (PAM clustering)
superclust <- cluster::pam(ok.som$codes[[1]], 2)
superclasses <- unname(superclust$clustering)
aweSOMscreeplot(ok.som, method = 'hierarchical', 
                hmethod = 'ward.D2', nclass = 2)
}
