% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_panel_functions.R
\name{aweSOMsilhouette}
\alias{aweSOMsilhouette}
\title{Silhouette plot}
\usage{
aweSOMsilhouette(ok.som, ok.sc)
}
\arguments{
\item{ok.som}{```kohonen``` object, a SOM created by the ```som``` function.}

\item{ok.sc}{Computed super-classes object resulting from either PAM or hierarchical clustering}
}
\value{
Silhouette plot for superclass clustering
}
\description{
Plots a silhouette plot that provides a quality measurement of the superclasses of the SOM. 
Available for for both PAM and hierarchical clustering.
}
\examples{
## Build training data
dat <- iris[, c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")]
### Scale training data
dat <- scale(dat)
## Train SOM
### RNG Seed (for reproducibility)
### Initialization (PCA grid)
init <- somInit(dat, 4, 4)
ok.som <- kohonen::som(dat, grid = kohonen::somgrid(4, 4, 'hexagonal'), 
                       rlen = 100, alpha = c(0.05, 0.01), 
                       radius = c(6.08,-6.08), init = init,
                       dist.fcts = 'sumofsquares')
## Group cells into superclasses (PAM clustering)
superclust <- cluster::pam(ok.som$codes[[1]], 2)
superclasses <- unname(superclust$clustering)
aweSOMsilhouette(ok.som, superclasses)
}
