% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_panel_functions.R
\name{aweSOMsmoothdist}
\alias{aweSOMsmoothdist}
\title{Smooth Distance Plot}
\usage{
aweSOMsmoothdist(x, pal = "viridis", reversePal = F)
}
\arguments{
\item{x}{SOM data object}

\item{pal}{The color palette for visualizing distance. Default is "viridis"}

\item{reversePal}{Boolean whether color palette for variables is reversed. Default is FALSE}
}
\value{
Smooth distance plot
}
\description{
TBD
}
\examples{
ok.data <- iris
## Build training data
dat <- ok.data[,c("Sepal.Length", "Sepal.Width",  "Petal.Length", "Petal.Width" )]
### Scale training data
dat <- scale(dat)
## Train SOM
### RNG Seed (for reproducibility)
### Initialization (PCA grid)
init <- aweSOM::somInit(dat, 4, 4)
ok.som <- kohonen::som(dat, grid = kohonen::somgrid(4, 4, 'hexagonal'), rlen = 100, alpha = c(0.05, 0.01), radius = c(6.08,-6.08), init = init, dist.fcts = 'sumofsquares')
aweSOM::aweSOMsmoothdist(ok.som)
}
