% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{somDist}
\alias{somDist}
\title{Distance measures of a SOM}
\usage{
somDist(som)
}
\arguments{
\item{som}{```kohonen``` object, a SOM created by the ```som``` function.}
}
\value{
A list with distance measures
}
\description{
Several distance measures between cells or prototypes of a trained SOM (in
grid space, in data space).
}
