% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{somQuality}
\alias{somQuality}
\title{Calculate SOM quality measures}
\usage{
somQuality(som, traindat)
}
\arguments{
\item{som}{`kohonen` object, a SOM created by the `kohonen::som` function.}

\item{traindat}{matrix containing the training data}
}
\value{
A list with quality measures : quantization error, share of explained
  variance, topographic error and Kaski-Lagus error.
}
\description{
Calculate SOM quality measures
}
